feat(p10,10M-1): parser+validator memory64 widening (MemoryEntry.idx_type, u64 limits)
Lands ADR-0111 Decision 1 (idx_type discriminator always present),
plus the i64 LEB widening for limits values. Parser cascade:

- MemoryEntry gains `idx_type: enum(u1) { i32, i64 } = .i32`;
  min/max widen `u32 → u64` (always — i32 memories store values
  ≤ u32::MAX; i64 memories use the full u64 range per spec).
- decodeMemory() reads Wasm 3.0 §5.4.4 4-bit limits prefix:
  accepts 0x00/0x01 always; 0x04/0x05 only when
  `comptime build_options.wasm_level >= .v3_0` (else
  Error.Memory64Unsupported per ADR-0111 D1); rejects shared
  (0x02 — threads OOS) and reserved bits.
- Multi-memory enable at parser layer: decodeMemory loop already
  supported count > 1; validator memory_count plumbing too. The
  count > 1 reject lives at instantiate.zig:582 (runtime layer
  — touched at 10.M-2).
- engine/compile.zig validator: per-idx_type page cap (i32 keeps
  65536; i64 uses 2^32 per Wasm 3.0 §A.1 implementation-limit
  ceiling).

Symmetric ImportPayload.memory + ImportShape.memory (Instance)
cascade — import-declared memory carries idx_type+u64 limits
through cross-module type-match (instantiate.zig:779) and host
linker (zwasm/linker.zig:263).

spec_assert_runner_{base,non_simd} cascade: helper return types
widened u32→u64; @intCast at the resetGrowableMemory call sites
(runner is Wasm 1.0/2.0-scope per ROADMAP §9.x — i64 fixtures
arrive at 10.M-5 with the runtime cascade in 10.M-2).

Tests: existing `decodeMemory: min only + min/max forms` adapted
for new IdxType. Added:

- decodeMemory: rejects shared flag (threads not supported)
- decodeMemory: rejects reserved flag bits
- decodeMemory: i64 memory min only (Wasm 3.0 §5.4.4)
- decodeMemory: i64 memory min+max (Wasm 3.0 §5.4.4)
- decodeMemory: multi-memory parser enable (Wasm 3.0 §5.4.6)
  with parse-time reject branch for v2_0 builds

Test gate: Mac `zig build test-all` GREEN; lint GREEN; zone_check
+ file_size_check exit 0.

Refs: §10 / 10.M-1; ADR-0111 D1/D6; survey at
private/notes/p10-10M-1-survey.md.

+++ test/spec/spec_assert_runner_base.zig
@@ -1590,7 +1590,7 @@ pub fn effectiveMemory0Min(
     allocator: std.mem.Allocator,
     importer_wasm: []const u8,
     registered: ?*const std.StringHashMapUnmanaged(RegisteredExporter),
-) u32 {
+) u64 {
     var temp_arena = std.heap.ArenaAllocator.init(allocator);
     defer temp_arena.deinit();
     const ta = temp_arena.allocator();
@@ -1623,7 +1623,7 @@ pub fn effectiveMemory0Max(
     allocator: std.mem.Allocator,
     importer_wasm: []const u8,
     registered: ?*const std.StringHashMapUnmanaged(RegisteredExporter),
-) ?u32 {
+) ?u64 {
     var temp_arena = std.heap.ArenaAllocator.init(allocator);
     defer temp_arena.deinit();
     const ta = temp_arena.allocator();
@@ -1654,7 +1654,7 @@ fn extractExporterMemoryMax(
     allocator: std.mem.Allocator,
     exporter_wasm: []const u8,
     export_name: []const u8,
-) ?u32 {
+) ?u64 {
     var temp_arena = std.heap.ArenaAllocator.init(allocator);
     defer temp_arena.deinit();
     const ta = temp_arena.allocator();
@@ -1692,7 +1692,7 @@ fn extractExporterMemoryMin(
     allocator: std.mem.Allocator,
     exporter_wasm: []const u8,
     export_name: []const u8,
-) ?u32 {
+) ?u64 {
     var temp_arena = std.heap.ArenaAllocator.init(allocator);
     defer temp_arena.deinit();
     const ta = temp_arena.allocator();
@@ -1884,7 +1884,7 @@ pub fn resetGrowableMemory(initial_pages: u32) void {
 /// module's own declaration — matters for `memory_size.wast`
 /// and `memory_grow.wast` where the max-pages cap gates whether
 /// a `memory.grow` succeeds vs returns -1.
-pub fn extractMemoryLimits(allocator: std.mem.Allocator, wasm_bytes: []const u8) struct { min: u32, max: ?u32 } {
+pub fn extractMemoryLimits(allocator: std.mem.Allocator, wasm_bytes: []const u8) struct { min: u64, max: ?u64 } {
     var module = zwasm.parse.parser.parse(allocator, wasm_bytes) catch return .{ .min = 0, .max = null };
     defer module.deinit(allocator);
     const sec = module.find(.memory) orelse return .{ .min = 0, .max = null };
@@ -2224,8 +2224,8 @@ fn crossModuleMemoryMismatch(
     allocator: std.mem.Allocator,
     module: *const zwasm.runtime.Module,
     mem_idx: u32,
-    want_min: u32,
-    want_max: ?u32,
+    want_min: u64,
+    want_max: ?u64,
 ) bool {
     var imported_mems: u32 = 0;
     if (module.find(.import)) |is| {
@@ -2249,7 +2249,7 @@ fn crossModuleMemoryMismatch(
     return memLimitsMismatch(m.min, m.max, want_min, want_max);
 }
 
-fn memLimitsMismatch(src_min: u32, src_max: ?u32, want_min: u32, want_max: ?u32) bool {
+fn memLimitsMismatch(src_min: u64, src_max: ?u64, want_min: u64, want_max: ?u64) bool {
     if (src_min < want_min) return true;
     if (want_max) |wm| {
         const sm = src_max orelse return true;
